+++ src/modules/searchBar/tests/unit/components/SearchBar.spec.js
@@ -24,7 +24,8 @@ describe("src/modules/searchBar/components/SearchBar.vue", () => {
         setSearchResultsActiveSpy,
         setShowAllResultsSpy,
         setShowSearchResultsInTreeSpy,
-        setCurrentActionEventSpy;
+        setCurrentActionEventSpy,
+        filterInLayerSelection = false;
 
 
     beforeEach(() => {
@@ -111,9 +112,16 @@ describe("src/modules/searchBar/components/SearchBar.vue", () => {
                                 showAllResults: () => false,
                                 suggestionListLength: () => 0,
                                 type: () => "searchBar",
-                                addLayerButtonSearchActive: () => true
+                                addLayerButtonSearchActive: () => true,
+                                searchIsLoading: () => false
                             },
                             mutations: searchBarMutationsSpy
+                        },
+                        LayerSelection: {
+                            namespaced: true,
+                            mutations: {
+                                setHighlightLayerId: sinon.stub()
+                            }
                         }
                     }
                 },
@@ -121,7 +129,8 @@ describe("src/modules/searchBar/components/SearchBar.vue", () => {
                     namespaced: true,
                     getters: {
                         titleBySide: () => () => true,
-                        currentComponent: () => () => "root"
+                        currentComponent: () => () => "root",
+                        previousNavigationEntryText: () => () => "common:modules.searchBar.search"
                     },
                     actions: menuActionsSpy
                 },
@@ -133,7 +142,8 @@ describe("src/modules/searchBar/components/SearchBar.vue", () => {
                 }
             },
             getters: {
-                portalConfig: sinon.stub()
+                portalConfig: sinon.stub(),
+                filterInLayerSelection: () => filterInLayerSelection
             },
             actions: {
                 initializeModule: sinon.stub()
@@ -143,6 +153,7 @@ describe("src/modules/searchBar/components/SearchBar.vue", () => {
 
     afterEach(() => {
         sinon.restore();
+        filterInLayerSelection = false;
     });
 
     describe("render SearchBar", () => {
@@ -296,6 +307,8 @@ describe("src/modules/searchBar/components/SearchBar.vue", () => {
 
     describe("checkCurrentComponent ", () => {
         it("startSearch is executed for modules", async () => {
+            filterInLayerSelection = true;
+
             wrapper = await mount(SearchBarComponent, {
                 global: {
                     plugins: [store]
@@ -307,7 +320,7 @@ describe("src/modules/searchBar/components/SearchBar.vue", () => {
             expect(startSearchSpy.called).to.be.true;
         });
         it("startSearch is executed for layerSelection", async () => {
-            searchInputValue = "";
+            searchInputValue = "valid-search";
             wrapper = await mount(SearchBarComponent, {
                 global: {
                     plugins: [store]
@@ -317,10 +330,25 @@ describe("src/modules/searchBar/components/SearchBar.vue", () => {
 
             wrapper.vm.checkCurrentComponent("layerSelection");
 
-
             expect(startSearchSpy.called).to.be.true;
-            expect(menuActionsSpy.navigateBack.called).to.be.true;
             expect(searchBarActionsSpy.startLayerSelectionSearch.called).to.be.true;
+        });
+        it("navigateBack is called for layerSelection with empty searchInput", async () => {
+            searchInputValue = "";
+
+            wrapper = await mount(SearchBarComponent, {
+                global: {
+                    plugins: [store]
+                }
+            });
+
+            const navigateBackSpy = sinon.spy(wrapper.vm, "navigateBack");
+
+            wrapper.vm.checkCurrentComponent("layerSelection");
+
+            expect(navigateBackSpy.calledWith("mainMenu")).to.be.true;
+            expect(menuActionsSpy.navigateBack.called).to.be.true;
+            expect(searchBarActionsSpy.startLayerSelectionSearch.called).to.be.false;
         });
     });